Move ABC definition into main MPC file
